Changed location of essay source files
The location of the essay files is now hard-coded to be in
"http://lairbob.com/essays-html/"

Obviously, a better approach would be to at least infer the name of the
server, and use a fixed directory path from there. The _best_ approach
would be to make the parent path a setting, from within the plug-in's UI
admin panel.